moved enums for compatability into one place
