+++ Sources/WishKit/Extensions/Enum+Compat.swift
@@ -1,0 +1,22 @@
+//
+//  Enum+Compatability.swift
+//  wishkit-ios
+//
+//  Created by Martin Lasek on 9/28/23.
+//  Copyright © 2023 Martin Lasek. All rights reserved.
+//
+
+enum Compatability {
+    enum Visibility {
+        case automatic
+        case visible
+        case hidden
+    }
+
+    enum ControlSize {
+        case mini
+        case small
+        case regular
+        case large
+    }
+}
